+++ changelog.md
@@ -9,6 +9,8 @@
   - Introduced a new `accessoryView` prop that enables rendering any view on the leading side of the text input.
 - react-native-bpk-component-phone-input:
  - New `BpkDiallingCodeList` component.
+- react-native-bpk-component-badge:
+  - Introducing the new React Native badge component.
 
 ## 2018-03-01 - Fix for Android button props